[QDP-86] Improve tasks reporting mechanism

+++ qdp-core/src/main/java/com/quantori/qdp/core/task/model/ResultAggregator.java
@@ -5,7 +5,7 @@ public interface ResultAggregator {
 
     StreamTaskResult getResult();
 
-    float getPercent();
+    double getPercent();
 
     default void close() { }
